fix(ui): use avatar component instead of the gradient hook pattern (#2698)
1. Created a new `Avatar` component for the sidebar.
2. Delete the useAvatarGradient hook.
3. Renamed `profileimageurl` to `image_url` to match the latest value
coming from the API

This should perform better and be easier to manage going forward. 

<img width="1258" height="600" alt="CleanShot 2025-08-11 at 15 03 08@2x"
src="https://github.com/user-attachments/assets/abb12c8d-0342-4e2b-a9dd-3e0e8a084189"
/>
