+++ src/containers/main/Airdrop/sidebar/components/Avatar/Avatar.tsx
@@ -1,0 +1,47 @@
+import { AvatarWrapper } from './styles';
+
+type Seed = number;
+
+const hashString = (str: string): Seed => {
+    let hash: Seed = 0;
+    for (let i = 0; i < str.length; i++) {
+        const char: number = str.charCodeAt(i);
+        hash = (hash << 5) - hash + char;
+        hash = hash & hash;
+    }
+    return hash;
+};
+
+const seededRandom = (seed: Seed): number => {
+    const x: number = Math.sin(seed++) * 10000;
+    return x - Math.floor(x);
+};
+
+const generateColor = (seed: Seed): string => {
+    const r: number = Math.floor(seededRandom(seed) * 256);
+    const g: number = Math.floor(seededRandom(seed + 1) * 256);
+    const b: number = Math.floor(seededRandom(seed + 2) * 256);
+    return `rgb(${r}, ${g}, ${b})`;
+};
+
+interface Props {
+    username?: string;
+    image?: string;
+}
+
+export default function Avatar({ username, image }: Props) {
+    const finalUsername = username ? `@${username}` : '';
+    let finalImage = `url(${image})`;
+
+    if (!image && username) {
+        const seed = hashString(username);
+        const color1 = generateColor(seed);
+        const color2 = generateColor(seed + 3);
+        const xPos = Math.floor(seededRandom(seed + 6) * 100);
+        const yPos = Math.floor(seededRandom(seed + 7) * 100);
+
+        finalImage = `radial-gradient(circle at ${xPos}% ${yPos}%, ${color1}, ${color2})`;
+    }
+
+    return <AvatarWrapper $image={finalImage} aria-label={finalUsername} title={finalUsername} />;
+}
